pyproject.toml
[project]
name = "nanobot-ai"
version = "0.1.3.post4"
description = "A lightweight personal AI assistant framework"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "nanobot contributors"}
]
keywords = ["ai", "agent", "chatbot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "typer>=0.9.0",
    "litellm>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "fastapi>=0.115.0",
    "websockets>=12.0",
    "websocket-client>=1.6.0",
    "httpx>=0.25.0",
    "loguru>=0.7.0",
    "paho-mqtt>=2.1.0",
    "readability-lxml>=0.8.0",
    "rich>=13.0.0",
    "croniter>=2.0.0",
    "python-telegram-bot>=21.0",
    "uvicorn>=0.32.0",
]

[project.optional-dependencies]
feishu = [
    "lark-oapi>=1.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[project.scripts]
nanobot = "nanobot.cli.commands:app"
hatchery-service = "hatchery.service_runner:main"
hatchery-bridge = "hatchery.orchestrator_bridge.runner:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["nanobot", "hatchery"]

[tool.hatch.build.targets.wheel.sources]
"nanobot" = "nanobot"
"hatchery" = "hatchery"

[tool.hatch.build]
include = [
    "nanobot/**/*.py",
    "nanobot/skills/**/*.md",
    "nanobot/skills/**/*.sh",
    "hatchery/**/*.py",
    "hatchery/**/*.html",
]

[tool.hatch.build.targets.sdist]
include = [
    "nanobot/",
    "hatchery/",
    "bridge/",
    "README.md",
    "LICENSE",
]

[tool.hatch.build.targets.wheel.force-include]
"bridge" = "nanobot/bridge"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]